api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic
python-dotenv
loguru
alembic
pydantic[email]
argon2-cffi
passlib[bcrypt]
python-jose[cryptography]
requests
python-multipart
pypdf
scikit-learn
spacy
nltk
stripe
apscheduler
beautifulsoup4
pytest
